* Move scalar[] to public entry struct.  As a result, dirfile_add and dirfile_alter_entry can be used to add/modify fields with non-literal parameters. * Make shift parameter for PHASE long. * Use unsigned long consistently for dirfile flags/encoding/byte_sec * Internally assume int type parameters are 16-bits wide (ie. spf, bitnum, numbits). * Use _gd_spf_t internally to keep track of SPF type. * Added GD_PRETTY_PRINT and some rudimentary prettification of format files. * Bumped DIRFILE_STANDARDS_VERSION. * [BUG]: Clear dirfile error on success of dirfilename(), dirfile_reference(), get_reference(). * More man page updates.

+++ bindings/cxx/dirfile.cpp
@@ -1,4 +1,4 @@
-// (C) 2008 D. V. Wiebe
+// (C) 2008, 2009 D. V. Wiebe
 //
 ///////////////////////////////////////////////////////////////////////////
 //
@@ -28,7 +28,7 @@ Dirfile::Dirfile()
   error_string = NULL;
 }
 
-Dirfile::Dirfile(const char* filedir, unsigned int flags,
+Dirfile::Dirfile(const char* filedir, unsigned long flags,
     gd_parser_callback_t sehandler, void* extra)
 {
   D = dirfile_cbopen(filedir, flags, sehandler, extra);
@@ -134,7 +134,7 @@ const char* Dirfile::ErrorString(size_t len)
   return get_error_string(D, error_string, len);
 }
 
-int Dirfile::Include(const char* file, int format_file, unsigned int flags)
+int Dirfile::Include(const char* file, int format_file, unsigned long flags)
 {
   return dirfile_include(D, file, format_file, flags);
 }
